+++ CPU/ST/CPUSingleThreadLZSS.cpp
@@ -13,18 +13,18 @@
 #include "../BlockHelper.h"
 #include "CPUSingleThreadLZSS.h"
 
-std::pair<bool, double> CPUSingleThreadLZSS::compress(const uint8_t* inBuf, int64_t inSize,
-    uint8_t* outBuf, int64_t& outSize,
-    CompressFlagBlock* flagOut, int64_t nFlagBlocks, int64_t& flagSize)
+std::pair<bool, double> CPUSingleThreadLZSS::compress(const uint8_t* inBuf, int inSize,
+    uint8_t* outBuf, int& outSize,
+    CompressFlagBlock* flagOut, int nFlagBlocks, int& flagSize)
 {
     Timer timer;
 
     outSize = flagSize = 0;
-    for (int64_t i = 0; i < nFlagBlocks; ++i) {
+    for (int i = 0; i < nFlagBlocks; ++i) {
         BlockCompress(i, inBuf, inSize, outBuf, outSize, flagOut, flagSize,
-            [nFlagBlocks](int64_t blockId) {
+            [nFlagBlocks](int blockId) {
                 if ((blockId + 1) % 100 == 0) {
-                    printf("Block %" PRId64 "/%" PRId64 " done.\n", blockId + 1, nFlagBlocks);
+                    printf("Block %d/%d done.\n", blockId + 1, nFlagBlocks);
                 }
             });
     }
@@ -32,12 +32,12 @@ std::pair<bool, double> CPUSingleThreadLZSS::compress(const uint8_t* inBuf, int6
     return std::make_pair(true, timer.end());
 }
 
-std::pair<bool, double> CPUSingleThreadLZSS::decompress(CompressFlagBlock* flagIn, int64_t nFlagBlocks,
-    const uint8_t* inBuf, int64_t inSize, uint8_t* outBuf, int64_t outSize)
+std::pair<bool, double> CPUSingleThreadLZSS::decompress(CompressFlagBlock* flagIn, int nFlagBlocks,
+    const uint8_t* inBuf, int inSize, uint8_t* outBuf, int outSize)
 {
     Timer timer;
 
-    for (int64_t i = 0; i < nFlagBlocks; ++i) {
+    for (int i = 0; i < nFlagBlocks; ++i) {
         BlockDecompress(i, flagIn, inBuf, outBuf);
     }
